LUSIA Studio - Backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
supabase==2.3.4
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.5
httpx==0.25.2
python-multipart==0.0.6
arq>=0.26.1
redis>=5.0.0
mistralai>=1.0.0
pypandoc>=1.14
